requirement.txt
absl-py==2.3.0
acres==0.5.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.45
botocore==1.38.45
certifi==2025.6.15
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
cloudinary==1.44.1
configobj==5.0.9
configparser==7.2.0
deepface==0.0.93
dotenv==0.9.9
etelemetry==0.3.1
fastapi==0.115.14
filelock==3.18.0
fire==0.7.0
fitz==0.0.1.dev2
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
frontend==0.0.3
fsspec==2025.5.1
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.73.1
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
httplib2==0.22.0
huggingface-hub==0.33.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
keras==3.10.0
libclang==18.1.1
looseversion==1.3.0
lxml==6.0.0
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.1.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
pathlib==1.0.1
pillow==11.2.1
protobuf==5.29.5
prov==2.1.1
psycopg2-binary==2.9.10
puremagic==1.29
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.26.1
pyparsing==3.2.3
PySocks==1.7.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
regex==2024.11.6
requests==2.32.4
retina-face==0.0.17
rich==14.0.0
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==4.1.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
starlette==0.46.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
traits==7.0.2
transformers==4.53.0
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
underthesea==6.8.4
underthesea_core==1.0.4
urllib3==2.5.0
uvicorn==0.34.3
Werkzeug==3.1.3
wrapt==1.17.2
